requirements.txt
pandas
requests
numpy
sklearn
matplotlib
python-dotenv
discord.py